Add DiscountCalculatorSpec

+++ src/main/java/pl/zycienakodach/pragmaticflights/modules/discounts/api/event/DiscountValueCalculated.java
@@ -1,4 +1,6 @@
 package pl.zycienakodach.pragmaticflights.modules.discounts.api.event;
 
-public record DiscountValueCalculated(String orderId, double discountInEuro) implements DiscountsEvent {
+import java.math.BigDecimal;
+
+public record DiscountValueCalculated(String orderId, BigDecimal discountInEuro) implements DiscountsEvent {
 }
