+++ src/main/java/pl/zycienakodach/pragmaticflights/modules/pricing/api/commands/ApplyOrderPriceDiscount.java
@@ -1,7 +1,9 @@
 package pl.zycienakodach.pragmaticflights.modules.pricing.api.commands;
 
+import java.math.BigDecimal;
+
 public record ApplyOrderPriceDiscount(
     String orderId,
-    double discountInEuro
+    BigDecimal discountInEuro
 ) {
 }
