+++ src/main/java/pl/zycienakodach/pragmaticflights/modules/pricing/api/events/CalculateOrderTotalPriceCompleted.java
@@ -1,6 +1,8 @@
 package pl.zycienakodach.pragmaticflights.modules.pricing.api.events;
 
+import java.math.BigDecimal;
+
 public record CalculateOrderTotalPriceCompleted(
-    String orderId, double regularPriceInEuro, double discountPriceInEuro, double totalPriceInEuro
+    String orderId, BigDecimal regularPriceInEuro, BigDecimal discountPriceInEuro, BigDecimal totalPriceInEuro
 ) implements PricingEvent {
 }
