+++ src/main/java/pl/zycienakodach/pragmaticflights/modules/pricing/api/events/CalculateOrderTotalPriceStarted.java
@@ -1,6 +1,8 @@
 package pl.zycienakodach.pragmaticflights.modules.pricing.api.events;
 
+import java.math.BigDecimal;
+
 public record CalculateOrderTotalPriceStarted(
-    String orderId, double regularPriceInEuro
+    String orderId, BigDecimal regularPriceInEuro
 ) implements PricingEvent {
 }
